refactor/ weather -> weather_provider, location_info -> location_provider
provider 이름을 통일성있게 유지하기 위해 이름을 변경
