+++ lib/pages/constitution_page.dart
@@ -1,7 +1,7 @@
 import 'dart:math';
 import 'package:sweater/providers/user_provider.dart';
 import 'package:sweater/providers/coordi_provider.dart';
-import 'package:sweater/providers/weather.dart';
+import 'package:sweater/providers/weather_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:sweater/widgets/check_menu.dart';
 import 'package:provider/provider.dart';
