+++ lib/pages/gender_change_page.dart
@@ -3,7 +3,7 @@ import 'package:sweater/widgets/check_menu.dart';
 import 'package:sweater/providers/coordi_provider.dart';
 import 'package:sweater/theme/global_theme.dart';
 import 'package:sweater/providers/user_provider.dart';
-import 'package:sweater/providers/weather.dart';
+import 'package:sweater/providers/weather_provider.dart';
 import 'package:provider/provider.dart';
 import 'package:sweater/module/gender.dart';
