+++ lib/pages/home_page.dart
@@ -6,8 +6,8 @@ import 'package:sweater/theme/global_theme.dart';
 import 'package:sweater/pages/gender_change_page.dart';
 import 'package:sweater/pages/manage_location_page.dart';
 import 'package:sweater/pages/constitution_page.dart';
-import 'package:sweater/providers/location_info.dart';
-import 'package:sweater/providers/weather.dart';
+import 'package:sweater/providers/location_provider.dart';
+import 'package:sweater/providers/weather_provider.dart';
 import 'package:sweater/providers/coordi_provider.dart';
 import 'package:provider/provider.dart';
 import 'package:sweater/theme/sweater_icons.dart';
