+++ lib/pages/manage_location_page.dart
@@ -3,11 +3,11 @@ import 'package:shared_preferences/shared_preferences.dart';
 import 'package:sweater/widgets/location_tile.dart';
 import 'package:sweater/widgets/check_menu.dart';
 import 'package:sweater/pages/add_location_page.dart';
-import 'package:sweater/providers/location_info.dart';
+import 'package:sweater/providers/location_provider.dart';
 import 'package:sweater/providers/coordi_provider.dart';
 import 'package:sweater/providers/user_provider.dart';
 import 'package:provider/provider.dart';
-import 'package:sweater/providers/weather.dart';
+import 'package:sweater/providers/weather_provider.dart';
 import 'package:sweater/theme/sweater_icons.dart';
 
 class ManageLocationPage extends StatefulWidget {
