+++ lib/widgets/coordi_section.dart
@@ -3,7 +3,7 @@ import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
 import 'package:sweater/providers/coordi_provider.dart';
 import 'package:sweater/widgets/change_coordi_button.dart';
-import 'package:sweater/providers/weather.dart';
+import 'package:sweater/providers/weather_provider.dart';
 import 'package:sweater/theme/global_theme.dart';
 import 'dart:ui';
 import 'package:sweater/theme/sweater_icons.dart';
