+++ lib/widgets/hourly_weather_section.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:sweater/widgets/hourly_weather_card.dart';
 import 'package:provider/provider.dart';
 import 'package:sweater/providers/coordi_provider.dart';
-import 'package:sweater/providers/weather.dart';
+import 'package:sweater/providers/weather_provider.dart';
 import 'package:sweater/providers/user_provider.dart';
 
 class NoGlowScrollBehavior extends ScrollBehavior {
